fix deployments cmd
