+++ internal/cmd/local/local_deployments.go
@@ -14,9 +14,14 @@ type DeploymentsCmd struct {
 	Restart string `help:"Deployment to restart."`
 }
 
-func (d *DeploymentsCmd) Run(ctx context.Context, telClient telemetry.Client, k8sClient k8s.Client) error {
+func (d *DeploymentsCmd) Run(ctx context.Context, telClient telemetry.Client, provider k8s.Provider) error {
 	ctx, span := trace.StartSpan(ctx, "local deployments")
 	defer span.End()
+
+	k8sClient, err := defaultK8s(provider.Kubeconfig, provider.Context)
+	if err != nil {
+		return err
+	}
 
 	spinner := &pterm.DefaultSpinner
 	if err := checkDocker(ctx, telClient, spinner); err != nil {
